Move SupportMapper to exoplayer package.
